etl-search-storage/src/requirements.txt
mysql-connector-python
elasticsearch
minio
